Convert output module type to {esnext}
The main reason why {nodenext} is used up to now is that, all packages
using this package that rely on cjs outputs somehow (e.g. for running
their jest tests) can work as expected, otherwise {esnext} is preferred.

Switch to {esnext}, considering the required changes to all the required
changes in all the dependent packages.

Related: #5