Refactor to a monorepo based architecture
Because of the benefits of a monorepo, refactor the project architecture.
All of the scanner packages will be moved to this repository, including
the scanner package itself (which exist in this repository).

This commit only includes the scanner package. All of the other packages
({address-extractor}, {observation-extractor} and {watcher-data-extractor})
will be added to this repo in future commits.

In addition:
- Introduce base configs for c8, eslint and typescript so that packages
  can share the configs and without any duplication
- Add missing {package-lock.json}, as it needs to be included in source
  control
- Remove {package-lock.json} from {.gitignore} and add {yarn.lock} instead
- Update {.gitlab-ci.yml} to handle packages inside monorepo correctly
- Include a {settings.json} in {.vscode} in the root of the project so
  that vscode users can work with eslint in the monorepo without trouble
- Convert some 4-space-indented files to 2-space