Remove all {.eslintrc.json} files in packages
Eslint can find its config in parent directory, or parent's parent and so
on. For this monorepo, there is an {.eslintrc.json} in the root of the
monorepo that can be used in packages, so the configs inside of the
packages is not needed anymore and can be safely removed. (Please note
that they are all exactly the same as the root {.eslintrc.json}.)