Update {tsconfig} of all packages to match monorepo capabilities
Because there is a base {tsconfig} in the root of the monorepo, all
packages can extend it. Update all {tsconfig} of the packages to extend
the root one and remove redundant configs. (It includes removing {allowJs},
too, although it's not included in base config. {allowJs} is false by
default an we can remove it safely.)

In addition, remove {lib} config from {tsconfig.base.json}, as the {esnext}
module type in the config covers it.

Another important change is the addition of project references to all of
current packages but {scanner}. All of them need {scanner} to be built
and because {scanner} is part of the monorepo, it's just a symlink to the
package folder and doesn't get built when someone runs {npm install}. So
adding {scanner} as in {references} of all other packages is crucial. It's
clear that all {package.json} of those packages are also updated to run
{tsc} with {--build} flag to build the {scanner} package. {rimraf dist}
is also removed from {build} script of packages, as it conflicts with how
typescript handles project references. {*.tsbuildinfo} is also added to
{.gitignore}, as it's get generated by enabling project references.

All this aside, another inevitable change is the need for updating how
jest handles packages. The problem is that jest cannot work with esm well
and crashes. All packages other than {scanner} should transpile output of
scanner to cjs. So transpile the scanner output (only) for tests of all
packages but {scanner} (using {js-with-babel} preset of {ts-jest}). A
babel config is also added so that the preset can generate cjs output. It
follows the other configurations of the monorepo: There is a root config
and all of packages extend it as needed. Also remove redundant {transform}
config option in jest config of packages other than {scanner}, as the
preset handles it automatically.

Also sort all the configurations files (excluding {extends} config which
is better to come first in the file).