Fix `blob` with wrong content in `getRosenData` of `ogmios.ts`
If there are some wrong data in `blob` (e.g. some of the keys are missing
or the value type is wrong), it will cause the extractor to crash.

Refactor how `getRosenData` is implemented so that it safely checks the
content of `blob` and prevents such crashes.

Also remove `getObjectKeyAsStringOrUndefined` and all its tests as it's
no longer needed.