fix(scanner): fix ogmios fork block handling
If a fork happens, ogmios sends us two blocks with the same height but
different hashes. In this situation, because the fork block parent is
not last saved block (because they are the same blocks but with
different hashes), scanning process gets stuck and gives us error.

Fix the issue by forking at the height of new incoming block. By doing
so, if the block is new, forking does nothing and the process continues.
If the block is a fork, it removes the last block and resumes the
scanning process. So in both scenarios, we don't get stuck anymore.