Update extractor initialization procedure.
Lately, we just initialized the extractors just once, when the scanner
starts to run. But they may have been initialized in a forked block, and
we ignore all spent boxes in initialization step. Then we might lose
some boxes that we thought are spent in the forked block.
Then after we stepped backward from forked blocks, we will check if the
extractors need to be reinitialized.