Store spend info batch and remove init round 2
With parallel network requests we can't ensure blocks to be processed in
order (Also node api sort is based on its index not the transaction
heights). Then we had to run the whole process twice.
To optimize this procedure we decided to store all spending information
at the first round and reapply all of them after the initialization
completes.