Optimize explorer initializer by storing range queries
A `RangeQuery` represents the number of records within a specified range
(from start to end). By storing range queries, we avoid the need to
repeatedly query the same range multiple times.
A `RangeList` is a collection of `RangeQuery` items that are queued to
manage and limit ranges. This allows us to focus on processing ranges
where the number of records is sufficiently small.