Update ergo scanner to use NetworkConnectorManager
Cardano scanner now uses NetworkConnectorManager to manage multiple
connectors. The abstract class is added to remove the duplicated
functionalities in cardano scanners.
The creation of netowrk manager is deligated to the scanner user and the
scanner is now uses the provided network to operate. All connectors
configuration are removed from its constructor and  the user is responsible
for providing the correct network to the scanner.
Cardano blockFrost, koios and graphql networks now are available as exported
classes.