Update bitcoin esplora scanner to use NetworkConnectorManager
Bitcoin esplora scanner now uses NetworkConnectorManager to manage multiple
connectors.
The creation of netowrk manager is deligated to the scanner user and the
scanner is now uses the provided network to operate. All connectors
configuration are removed from its constructor and  the user is responsible
for providing the correct network to the scanner.
Bitcoin esplora and doge esplora networks now are available as exported
classes.