Update evm rpc scanner to use NetworkConnectorManager
Evm rpc scanner now uses NetworkConnectorManager to manage multiple
connectors.
The creation of netowrk manager is deligated to the scanner user and the
scanner is now uses the provided network to operate. All connectors
configuration are removed from its constructor and  the user is responsible
for providing the correct network to the scanner.
Evm rpc network now is available as exported class.